docs: change mod display name to "The Fence Unleashed"
> The initial name for this mod was "Fency" (and this name is still
  used in most places) but it had to be changed for technical reasons.

The CurseForge project was created right as their servers were having issues and even though the project was reportedly created successfully, it never showed up in my Dashboard. Attempts to create it again after the servers were back to normal failed because the project name was already taken. A ticket about this was closed roughly a week later with a copy&paste answer about the difference between project names and slugs (Thanks for nothing, I guess?), so I decided to simply change the display name of the mod.

Should someone else actually own the project "Fency", I'll can just change the mod's ID and (and other references) later, but I doubt it...